Add files via upload
It was a University project for MultiMedia course
AudioPlayer_PictureEditor are two different projects and I've put them in one MDI with friendly GUI, The AudioPlayer project has GUI that will let the user to pick an audio play it and edit on it beside the ability to merge two audios together and save all the changes. The Picture Editor project has GUI that will let the user to pick a photo and edit the colors brightness or even draw on it and save all the changes too. You can open multi projects of both photo and audio project.